refactor: rename ListStoryComments to StoryCommentsTable
This commit refactors the `ListStoryComments` Livewire component to `StoryCommentsTable`.

This renaming provides a more accurate and descriptive name for the component, reflecting its role as a table for displaying story comments, and promotes better code readability and maintainability. It also includes renaming the view and tests related to this component, as well as updating references to the component in other views.
